Move Source and Layer back into structs
and make structs a single file again
